fixes
HOSTNAME -> HOSTUNAME so the user's hostname is used
Let the user choose if they want microcode updates or not

+++ arch-install-auto.sh
@@ -5,7 +5,7 @@ echo "Ensure you have a working internet connection."
 echo "This script overwrites all data on the selected disk."
 
 echo "Name the system: "
-read HOSTNAME
+read HOSTUNAME
 
 # Test UEFI BIOS
 EFIVARS="$(ls /sys/firmware/efi/efivars)"
@@ -60,32 +60,46 @@ hwclock --systohc
 echo "en_US.UTF-8 UTF-8" >> /etc/locale.gen
 locale-gen
 echo "LANG=en_US.UTF-8" > /etc/locale.conf
-echo "$HOSTNAME" > /etc/hostname
+echo "$HOSTUNAME" > /etc/hostname
 echo "127.0.0.1	localhost" >> /etc/hosts
 echo "::1	localhost" >> /etc/hosts
-echo "127.0.1.1	$HOSTNAME.localdomain $HOSTNAME" >> /etc/hosts
+echo "127.0.1.1	$HOSTUNAME.localdomain $HOSTUNAME" >> /etc/hosts
 echo "Setting root password to root."
 echo root:root | chpasswd
 
 BLOCKDEV=$(cat /root/blockdev.tmp)
 
+function cpuvendor() {
 echo "Detecting CPU Vendor..."
-if [[ "$(lscpu)" =~ "Intel" ]]; then
-	echo "Intel CPU detected."
-	pacman -S intel-ucode --noconfirm
-	MICROINITRD='intel-ucode.img'
-fi
-if [[ "$(lscpu)" =~ "AMD" ]]; then
-	echo "AMD CPU detected."
-	pacman -S amd-ucode --noconfirm
-	MICROINITRD='amd-ucode.img'
+	if [[ "$(lscpu)" =~ "Intel" ]]; then
+		echo "Intel CPU detected."
+		pacman -S intel-ucode --noconfirm
+		MICROINITRD='intel-ucode.img'
+	fi
+	if [[ "$(lscpu)" =~ "AMD" ]]; then
+		echo "AMD CPU detected."
+		pacman -S amd-ucode --noconfirm
+		MICROINITRD='amd-ucode.img'
+	fi
+}
+
+echo "Enable microcode updates? [y/n]"
+read microask
+
+if [ "$microask" = 'y' ]; then
+	cpuvendor
 fi
 
 ROOTUUID=$(blkid -o value -s PARTUUID "/dev/${BLOCKDEV}3")
 
 echo "Adding boot entry..."
-echo "efibootmgr --disk "/dev/${BLOCKDEV}" --part 1 --create --label "Arch Linux" --loader /vmlinuz-linux --unicode 'root=PARTUUID='"${ROOTUUID}"' rw initrd=\'"${MICROINITRD}"' initrd=\initramfs-linux.img' --verbose" | bash
-echo "efibootmgr --disk "/dev/${BLOCKDEV}" --part 1 --create --label "Arch Linux" --loader /vmlinuz-linux --unicode 'root=PARTUUID='"${ROOTUUID}"' rw initrd=\'"${MICROINITRD}"' initrd=\initramfs-linux.img' --verbose" > /root/efi-boot-vars
+if [ "$microask" = 'y' ]; then
+	echo "efibootmgr --disk "/dev/${BLOCKDEV}" --part 1 --create --label "Arch Linux" --loader /vmlinuz-linux --unicode 'root=PARTUUID='"${ROOTUUID}"' rw initrd=\'"${MICROINITRD}"' initrd=\initramfs-linux.img' --verbose" | bash
+	echo "efibootmgr --disk "/dev/${BLOCKDEV}" --part 1 --create --label "Arch Linux" --loader /vmlinuz-linux --unicode 'root=PARTUUID='"${ROOTUUID}"' rw initrd=\'"${MICROINITRD}"' initrd=\initramfs-linux.img' --verbose" > /root/efi-boot-vars
+else
+	echo "efibootmgr --disk "/dev/${BLOCKDEV}" --part 1 --create --label "Arch Linux" --loader /vmlinuz-linux --unicode 'root=PARTUUID='"${ROOTUUID}"' rw initrd=\initramfs-linux.img' --verbose" | bash
+	echo "efibootmgr --disk "/dev/${BLOCKDEV}" --part 1 --create --label "Arch Linux" --loader /vmlinuz-linux --unicode 'root=PARTUUID='"${ROOTUUID}"' rw initrd=\initramfs-linux.img' --verbose" > /root/efi-boot-vars
+fi
 echo "Install complete, enable and/or disable network profiles as necessary."
 echo "Consider running the post-install script after rebooting."
 echo "NOTE: root password is 'root'!"
